pyproject.toml
[tool.poetry]
name = "qparcchallenge2022"
version = "0.1.0"
description = ""
authors = ["kotamanegi <kotamanegi84@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyscf = "^2.0.1"
openfermionpyscf = "^0.5"
Qulacs = "^0.3.0"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.950"
prospector = "^1.7.7"
pylama = "^8.3.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
